+++ hnix-store-core/src/System/Nix/Nar.hs
@@ -6,10 +6,8 @@ Maintainer  : Shea Levy <shea@shealevy.com>
 {-# language ScopedTypeVariables #-}
 {-# language TypeFamilies        #-}
 
-
 module System.Nix.Nar
   (
-
   -- * Encoding and Decoding NAR archives
     buildNarIO
   , unpackNarIO
@@ -43,15 +41,13 @@ import qualified Control.Concurrent                as Concurrent
 import qualified Data.ByteString                   as BS
 import qualified System.IO                         as IO
 
-import qualified System.Nix.Internal.Nar.Effects   as Nar
-import qualified System.Nix.Internal.Nar.Options   as Nar
-import qualified System.Nix.Internal.Nar.Parser    as Nar
-import qualified System.Nix.Internal.Nar.Streamer  as Nar
-
+import qualified System.Nix.Nar.Effects   as Nar
+import qualified System.Nix.Nar.Options   as Nar
+import qualified System.Nix.Nar.Parser    as Nar
+import qualified System.Nix.Nar.Streamer  as Nar
 
 -- For a description of the NAR format, see Eelco's thesis
 -- https://nixos.org/~eelco/pubs/phd-thesis.pdf
-
 
 -- | Pack the filesystem object at @FilePath@ into a NAR and stream it into the
 --   @IO.Handle@
@@ -66,7 +62,6 @@ buildNarIO effs basePath outHandle =
     effs
     basePath
     (\chunk -> BS.hPut outHandle chunk >> Concurrent.threadDelay 10)
-
 
 -- | Read NAR formatted bytes from the @IO.Handle@ and unpack them into
 --   file system object(s) at the supplied @FilePath@